requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
ecdsa==0.19.1
fastapi==0.133.0
greenlet==3.3.2
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
# Editable Git install with no remote (qbem-backoffice-eb==0.1.0)
-e c:\users\user\desktop\dados_qbem_v3\services\qbem_backoffice_eb
rsa==4.9.1
shareds @ git+https://github.com/qbem-repos/dados-py-shareds.git@3f631dda8328e511b8d3b6cddc40c2c009557085
six==1.17.0
SQLAlchemy==2.0.47
SQLAlchemy-Utils==0.42.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
zxcvbn==4.5.0
